requirements.txt
selenium==4.9.1
python-dotenv==1.0.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.86.0
webdriver-manager==3.8.6
